Przeniesienie wszystkich klas akcji do osobnego folderu
