+++ src/main/java/org/example/Exam.java
@@ -1,6 +1,8 @@
 package org.example;
 
-public class Exam implements Event{
+import org.example.events.Event;
+
+public class Exam implements Event {
     @Override
     public void modifyStudent(Student student) {
