--- src/main/java/org/example/Party.java
@@ -1,11 +1,0 @@
-package org.example;
-
-public class Party implements Event{
-    @Override
-    public void modifyStudent(Student student) {
-       // student.increaseHealth(10);
-             int money=student.getPredisposition("Wealth")/2;
-        student.setPredisposition("Wealth", money);
-
-    }
-}
