--- src/main/java/org/example/Rest.java
@@ -1,8 +1,0 @@
-package org.example;
-
-public class Rest implements Event{
-    @Override
-    public void modifyStudent(Student student) {
-        student.increaseHealth(10);
-    }
-}
