+++ src/main/java/org/example/Student.java
@@ -1,7 +1,6 @@
 package org.example;
 
 
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.IntStream;
