--- src/main/java/org/example/Study.java
@@ -1,9 +1,0 @@
-package org.example;
-
-public abstract class Study implements Event{
-    @Override
-    public void modifyStudent(Student student){
-
-    }
-
-}
