+++ src/main/java/org/example/events/Event.java
@@ -1,4 +1,6 @@
-package org.example;
+package org.example.events;
+
+import org.example.Student;
 
 public interface Event {
     void modifyStudent(Student student);
