+++ src/main/java/org/example/events/Rest.java
@@ -1,0 +1,11 @@
+package org.example.events;
+import org.example.Student;
+
+import java.util.Random;
+
+public class Rest implements Event {
+    @Override
+    public void modifyStudent(Student student) {
+        student.increaseHealth(new Random().nextInt(40)+20);
+    }
+}
