+++ src/main/java/org/example/events/Study.java
@@ -1,0 +1,8 @@
+package org.example.events;
+
+import org.example.Student;
+
+public abstract class Study implements Event {
+    @Override
+    public void modifyStudent(Student student){}
+}
