+++ src/main/java/org/example/events/StudyIT.java
@@ -1,6 +1,8 @@
-package org.example;
+package org.example.events;
 
-public class StudyIT extends Study{
+import org.example.Student;
+
+public class StudyIT extends Study {
     @Override
     public void modifyStudent(Student student) {
         student.increaseSkill("IT");
