+++ src/main/java/org/example/events/StudyMath.java
@@ -1,6 +1,8 @@
-package org.example;
+package org.example.events;
 
-public class StudyMath extends Study{
+import org.example.Student;
+
+public class StudyMath extends Study {
     @Override
     public void modifyStudent(Student student) {
         student.increaseSkill("Math");
